WINDUP-1683 Tech stats boxes report WIP, move the report files to rules.generation.techreport package
